requirements.txt
allure-pytest~=2.8.19
allure-python-commons~=2.8.19
apipkg~=1.5
attrs~=20.2.0
certifi~=2020.6.20
chardet~=3.0.4
execnet~=1.7.1
idna~=2.10
iniconfig~=1.1.1
numpy~=1.19.4
packaging~=20.4
pika~=1.1.0
pluggy~=0.13.1
port-for~=0.4
psycopg2-binary~=2.8.6
py~=1.9.0
pyparsing~=2.4.7
pytest~=6.2.5
pytest-forked~=1.3.0
pytest-xdist~=3.2.0
pytest-check~=2.1.4
pytest-instafail~=0.5.0
pytest-rerunfailures~=9.1.1
requests~=2.32.1
six~=1.15.0
testit-adapter-pytest==3.2.6
testit-api-client==4.0.0
testit-python-commons==3.2.6
toml~=0.10.2
urllib3~=1.26.18
Pillow==8.2.0
XlsxWriter~=3.0.3
PyPDF2~=2.11.1
opencv-python~=4.6.0.66
pdf2image~=1.16.0
pandas~=1.4.4
mariadb~=1.1.4
openpyxl~=3.0.10
jsonschema~=4.17.3
pixelmatch==0.3.0
boto3~=1.28.4
pymongo==4.4.1
zeep==4.2.1
lxml==4.9.3
pyzbar==0.1.9
xmltodict~=0.13.0
elasticsearch~=8.13.1
pytz~=2024.1
prometheus-client~=0.20.0
ruff==0.8.1
python-dateutil~=2.9.0.post0
xlrd~=2.0.1
filelock==3.16.1